refactor(Channels): fix incorrectly shared properties (#6262)
